rafactoring and splitting pickupmodule into more partials to reduce the number of methods in each partial
