+++ PickupBot.Commands/Infrastructure/Helpers/PickupHelpers.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Discord;
 using Discord.WebSocket;
@@ -16,7 +14,18 @@ namespace PickupBot.Commands.Infrastructure.Helpers
             return (IVoiceChannel)guild.VoiceChannels.FirstOrDefault(c => c.Name.Equals(name, StringComparison.OrdinalIgnoreCase))
                    ?? await guild.CreateVoiceChannelAsync(name, properties => properties.CategoryId = categoryId);
         }
-        
+
+        public static bool IsInPickupChannel(IChannel channel) => channel.Name.StartsWith("pickup", StringComparison.OrdinalIgnoreCase);
+
+        public static async Task<ITextChannel> GetPickupQueuesChannel(SocketGuild guild)
+        {
+            var queuesChannel = (ITextChannel)guild.TextChannels.FirstOrDefault(c =>
+                                    c.Name.Equals("active-pickups", StringComparison.OrdinalIgnoreCase)) ??
+                                await guild.CreateTextChannelAsync("active-pickups",
+                                    properties => { properties.Topic = "Active pickups, use reactions to signup"; });
+            return queuesChannel;
+        }
+
         public static string ParseSubscribers(PickupQueue queue)
         {
             var subscribers = queue.Subscribers.Select(w => w.Name).ToList();
@@ -24,7 +33,36 @@ namespace PickupBot.Commands.Infrastructure.Helpers
                 subscribers.AddRange(Enumerable.Repeat("[?]", (queue.MaxInQueue) - queue.Subscribers.Count));
 
             //if queue found and user is in queue
-            return String.Join(", ", subscribers);
+            return string.Join(", ", subscribers);
+        }
+
+        public static async Task NotifyUsers(PickupQueue queue, string serverName, IUser guildUser, params SocketGuildUser[] users)
+        {
+            var usersList = String.Join(Environment.NewLine, queue.Subscribers.Where(u => u.Id != guildUser.Id).Select(u => $@"  - {u.Name}"));
+            var header = $"**Contact your teammates on the \"{serverName}\" server and glhf!**";
+            var remember = $"**Remember** {Environment.NewLine}" +
+                           $"Remember to do `!leave {queue.Name}` if/when you leave the game to make room for those in the waiting list!";
+
+            var embed = new EmbedBuilder
+            {
+                Title = $"Queue {queue.Name} is ready to go!",
+                Description = $@"{header}{Environment.NewLine}{usersList}{Environment.NewLine}{remember}",
+                Footer = new EmbedFooterBuilder { Text = $"Provided by pickup-bot - {serverName}" },
+                Color = Color.Orange
+            }.Build();
+
+            foreach (var user in users)
+            {
+                try
+                {
+                    await user.SendMessageAsync(embed: embed);
+                    await Task.Delay(500);
+                }
+                catch (Exception)
+                {
+                    //_logger.LogError($"Failed to send DM to {PickupHelpers.GetNickname(user)}", ex);
+                }
+            }
         }
 
         public static string GetNickname(IUser user) =>
