+++ PickupBot.Commands/Modules/PickupModuleSubscriberActions.cs
@@ -18,7 +18,7 @@ namespace PickupBot.Commands.Modules
         [Summary("Take a spot in a pickup queue, if the queue is full you are placed on the waiting list.")]
         public async Task Add([Name("Queue name"), Summary("Queue name"), Remainder]string queueName = "")
         {
-            if (!IsInPickupChannel((IGuildChannel)Context.Channel))
+            if (!PickupHelpers.IsInPickupChannel((IGuildChannel)Context.Channel))
                 return;
 
             //find queue with name {queueName}
@@ -30,7 +30,7 @@ namespace PickupBot.Commands.Modules
         [Summary("Leave a queue, freeing up a spot.")]
         public async Task Remove([Name("Queue name"), Summary("Optional, if empty the !clear command will be used."), Remainder] string queueName = "")
         {
-            if (!IsInPickupChannel((IGuildChannel)Context.Channel))
+            if (!PickupHelpers.IsInPickupChannel((IGuildChannel)Context.Channel))
                 return;
 
             if (string.IsNullOrWhiteSpace(queueName))
@@ -54,7 +54,7 @@ namespace PickupBot.Commands.Modules
         [Summary("Leave all queues you have subscribed to, including waiting lists")]
         public async Task Clear()
         {
-            if (!IsInPickupChannel((IGuildChannel)Context.Channel))
+            if (!PickupHelpers.IsInPickupChannel((IGuildChannel)Context.Channel))
                 return;
 
             //find queues with user in it
@@ -89,7 +89,7 @@ namespace PickupBot.Commands.Modules
         [Summary("Subscribes or unsubscribes the user to the promote role to get notifications when queues are created of when the !promote command is used")]
         public async Task Subscribe()
         {
-            if (!IsInPickupChannel((IGuildChannel)Context.Channel))
+            if (!PickupHelpers.IsInPickupChannel((IGuildChannel)Context.Channel))
                 return;
 
             var role = Context.Guild.Roles.FirstOrDefault(w => w.Name == "pickup-promote") ??
@@ -109,142 +109,6 @@ namespace PickupBot.Commands.Modules
                 await user.AddRoleAsync(role);
                 await ReplyAsync($"{PickupHelpers.GetMention(user)} - you are now subscribed to get notifications on `!promote`").AutoRemoveMessage(10);
             }
-        }
-
-        private async Task<PickupQueue> LeaveInternal(PickupQueue queue, ISocketMessageChannel channel, IGuildUser user, bool notify = true)
-        {
-            var guild = (SocketGuild)user.Guild;
-            var subscriber = queue.Subscribers.FirstOrDefault(s => s.Id == user.Id);
-            if (subscriber != null)
-            {
-                queue.Updated = DateTime.UtcNow;
-                queue.Subscribers.Remove(subscriber);
-
-                await MoveUserFromWaitingListToSubscribers(queue, subscriber, channel, guild);
-
-                if (!queue.Subscribers.Any() && !queue.WaitingList.Any())
-                {
-                    notify = await DeleteEmptyQueue(queue, guild, channel, notify);
-                }
-                else
-                {
-                    await SaveStaticQueueMessage(queue, guild);
-                    await _queueRepository.UpdateQueue(queue);
-                }
-            }
-
-            if (notify)
-                await channel.SendMessageAsync($"`{queue.Name} - {PickupHelpers.ParseSubscribers(queue)}`").AutoRemoveMessage(10);
-
-            return queue;
-        }
-
-        private async Task MoveUserFromWaitingListToSubscribers(PickupQueue queue, Subscriber subscriber, ISocketMessageChannel channel, SocketGuild guild)
-        {
-            if (queue.WaitingList.Any())
-            {
-                var next = queue.WaitingList.First();
-                queue.WaitingList.RemoveAt(0);
-
-                queue.Subscribers.Add(next);
-
-                var nextUser = guild.GetUser(next.Id);
-                if (nextUser != null)
-                {
-                    await channel.SendMessageAsync($"{PickupHelpers.GetMention(nextUser)} - you have been added to '{queue.Name}' since {subscriber.Name} has left.").AutoRemoveMessage();
-                    if (queue.Started)
-                    {
-                        var team = queue.Teams.FirstOrDefault(w => w.Subscribers.Exists(s => s.Id == subscriber.Id));
-                        if (team != null)
-                        {
-                            team.Subscribers.Remove(team.Subscribers.Find(s => s.Id == subscriber.Id));
-                            team.Subscribers.Add(new Subscriber { Name = PickupHelpers.GetNickname(nextUser), Id = nextUser.Id });
-                            await ReplyAsync($"{PickupHelpers.GetMention(nextUser)} - you are on the {team.Name}").AutoRemoveMessage();
-                        }
-                    }
-
-                    await NotifyUsers(queue, guild.Name, nextUser);
-                }
-            }
-        }
-
-        private async Task<bool> DeleteEmptyQueue(PickupQueue queue, SocketGuild guild, ISocketMessageChannel channel, bool notify)
-        {
-            var result = await _queueRepository.RemoveQueue(queue.Name, queue.GuildId); //Try to remove queue if its empty
-            if (result)
-            {
-                var queuesChannel = await GetPickupQueuesChannel(guild);
-                if (!string.IsNullOrEmpty(queue.StaticMessageId))
-                    await queuesChannel.DeleteMessageAsync(Convert.ToUInt64(queue.StaticMessageId));
-            }
-
-            if (!notify) return false;
-
-            await channel.SendMessageAsync($"`{queue.Name} has been removed since everyone left.`").AutoRemoveMessage(10);
-
-            return false;
-        }
-
-        private async Task AddInternal(string queueName, SocketGuild guild, ISocketMessageChannel channel, IGuildUser user)
-        {
-            PickupQueue queue;
-            if (!string.IsNullOrWhiteSpace(queueName))
-            {
-                queue = await _queueRepository.FindQueue(queueName, guild.Id.ToString());
-            }
-            else
-            {
-                var queues = (await _queueRepository.AllQueues(guild.Id.ToString())).OrderByDescending(w => w.Readiness).ToList();
-                queue = queues.Any(q => q.Readiness < 100) ? queues.FirstOrDefault(q => q.Readiness < 100) : queues.FirstOrDefault();
-            }
-
-            if (queue == null)
-            {
-                await channel.SendMessageAsync($"`Queue with the name '{queueName}' doesn't exists!`").AutoRemoveMessage(10);
-                return;
-            }
-
-            if (!await VerifyUserFlaggedStatus(user))
-                return;
-
-            if (queue.Subscribers.Any(w => w.Id == user.Id))
-            {
-                await channel.SendMessageAsync($"`{queue.Name} - {PickupHelpers.ParseSubscribers(queue)}`");
-                return;
-            }
-
-            var activity = await _activitiesRepository.Find(user);
-            activity.PickupAdd += 1;
-            
-            queue.Updated = DateTime.UtcNow;
-            queue.Subscribers.Add(new Subscriber { Id = user.Id, Name = PickupHelpers.GetNickname(user) });
-
-            if (queue.Subscribers.Count >= queue.MaxInQueue)
-            {
-                if (queue.WaitingList.All(w => w.Id != user.Id))
-                {
-                    queue = await SaveStaticQueueMessage(queue, guild);
-
-                    await channel.SendMessageAsync($"`You have been added to the '{queue.Name}' waiting list`").AutoRemoveMessage(10);
-                }
-                else
-                {
-                    await channel.SendMessageAsync($"`You are already on the '{queue.Name}' waiting list`").AutoRemoveMessage(10);
-                }
-            }
-            else
-            {
-                queue = await SaveStaticQueueMessage(queue, guild);
-
-                if (queue.Subscribers.Count == queue.MaxInQueue)
-                {
-                    await NotifyUsers(queue, guild.Name, user, queue.Subscribers.Select(s => guild.GetUser(s.Id)).Where(u => u != null).ToArray());
-                }
-
-                await channel.SendMessageAsync($"`{queue.Name} - {PickupHelpers.ParseSubscribers(queue)}`");
-            }
-            
-            await _queueRepository.UpdateQueue(queue);
         }
     }
 }
